+++ Sources/Model/GraphQL/HyperwalletTransferMethodConfigurationField.swift
@@ -33,16 +33,16 @@ public protocol HyperwalletTransferMethodConfigurationField {
 }
 
 final class TransferMethodConfigurationFieldResult: HyperwalletTransferMethodConfigurationField {
-    private let transferMethodUIConfigurations: [TransferMethodConfiguration]?
+    private let transferMethodUIConfigurations: [HyperwalletTransferMethodConfiguration]?
     private let country: HyperwalletCountry?
 
     /// Creates a new instance of the 'HyperwalletTransferMethodConfigurationField' based on the
     /// transfer method configuration result
     ///
     /// - Parameters:
-    ///   - transferMethodUIConfigurations: the GraphQL `[TransferMethodUIConfiguration]`
+    ///   - transferMethodUIConfigurations: the GraphQL `[HyperwalletTransferMethodConfiguration]`
     ///   - country: the GraphQL `HyperwalletCountry`
-    init(_ transferMethodUIConfigurations: [TransferMethodConfiguration]?, _ country: HyperwalletCountry?) {
+    init(_ transferMethodUIConfigurations: [HyperwalletTransferMethodConfiguration]?, _ country: HyperwalletCountry?) {
         self.transferMethodUIConfigurations = transferMethodUIConfigurations
         self.country = country
     }