+++ Tests/TransferMethod/HyperwalletTransferMethodTests.swift
@@ -54,7 +54,7 @@ class HyperwalletTransferMethodTests: XCTestCase {
         XCTAssertNotNil(linkNext?.href)
 
         // check the bank card
-        let bankCard = transferMethods?.data?.first { ($0.type ?? "")  == "BANK_CARD" }
+        let bankCard = transferMethods?.data?.first { ($0.type ?? "") == "BANK_CARD" }
 
         XCTAssertEqual(bankCard?.type, "BANK_CARD")
         XCTAssertEqual(bankCard?.token, "trm-00002")
@@ -72,7 +72,7 @@ class HyperwalletTransferMethodTests: XCTestCase {
         XCTAssertEqual(bankCard?.getField(cardBrandFieldName), "VISA")
 
         // check the bank account
-        let bankAccount = transferMethods?.data?.first { ($0.type ?? "")  == "BANK_ACCOUNT" }
+        let bankAccount = transferMethods?.data?.first { ($0.type ?? "") == "BANK_ACCOUNT" }
 
         XCTAssertEqual(bankAccount?.type, "BANK_ACCOUNT")
         XCTAssertEqual(bankAccount?.token, "trm-00001")